requirements.txt
django == 3.1
pillow
django-admin-interface
django-crispy-forms
django-rest-framework
social-auth-app-django